- Introduce Unit factory - Introduce Abstract Unit class (does nothing yet).
Goal is to get unit code in abstract classes, make them work so we can add a cool effect to the carry-all. ow, the shadow of that thing is still a bit low. This is because i am toying with the effect of a carry-all getting up/down. Like the Dune 2 clone; The Sleeper Has Awakened. Neat effect!